requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
Django==5.0.6
django-celery-beat==2.6.0
django-stubs==5.0.2
django-stubs-ext==5.0.2
django-timezone-field==6.1.0
djangorestframework==3.15.1
idna==3.7
inflection==0.5.1
jmespath==1.0.1
kombu==5.3.7
markdown-it-py==3.0.0
mdurl==0.1.2
mysqlclient==2.2.4
packaging==24.1
prompt_toolkit==3.0.47
Pygments==2.18.0
python-crontab==3.1.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.5
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
types-PyYAML==6.0.12.20240311
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
